Add Tradingview Order Lines (#32)
* Update Cancel and Place Order Functions

Updated function to cancel and place a new order. Requires 2 tx's until mango-client can be updated with the modify order tx.

* Completed handleModifyOrder

Completed handleModifyOrder and removed handlePlaceOrder as it is no longer required.

* Formatting and Context Updates

Added some additional context to the Modify order notifications, specifying modified orders as Limit orders, and updated USD formatting.

* Update yarn.lock

* Added toggles for OpenOrders initialization and OrderInProgress

Added an additional order line update on initialization of an OpenOrders account and added a toggle for orders in progress to prevent a moved order line from moving back to it's original price whilst a transaction for an order line is pending.